+++ src/main/java/com/revolvingmadness/sculk/language/builtins/classes/instances/EntityInstance.java
@@ -1,9 +1,13 @@
 package com.revolvingmadness.sculk.language.builtins.classes.instances;
 
+import com.revolvingmadness.sculk.accessors.EntityAccessor;
+import com.revolvingmadness.sculk.language.ErrorHolder;
 import com.revolvingmadness.sculk.language.builtins.classes.BuiltinClass;
 import com.revolvingmadness.sculk.language.builtins.classes.BuiltinType;
 import com.revolvingmadness.sculk.language.builtins.classes.types.EntityType;
+import com.revolvingmadness.sculk.language.builtins.classes.types.StringType;
 import net.minecraft.entity.Entity;
+import net.minecraft.nbt.NbtElement;
 
 import java.util.Objects;
 
@@ -13,6 +17,15 @@ public class EntityInstance extends BuiltinClass {
 
     public EntityInstance(Entity value) {
         this.value = value;
+    }
+
+    @Override
+    public void deleteIndex(BuiltinClass index) {
+        if (!index.instanceOf(new StringType())) {
+            throw ErrorHolder.cannotIndexTypeByType(this.getType(), index.getType());
+        }
+
+        ((EntityAccessor) this.value).sculk$deleteCustomData(index.toString());
     }
 
     @Override
@@ -28,6 +41,21 @@ public class EntityInstance extends BuiltinClass {
     }
 
     @Override
+    public BuiltinClass getIndex(BuiltinClass index) {
+        if (!index.instanceOf(new StringType())) {
+            throw ErrorHolder.cannotIndexTypeByType(this.getType(), index.getType());
+        }
+
+        NbtElement result = ((EntityAccessor) this.value).sculk$readCustomData(index.toString());
+
+        if (result == null) {
+            throw ErrorHolder.dictionaryHasNoKey(index.toString());
+        }
+
+        return BuiltinClass.fromNbtElement(result);
+    }
+
+    @Override
     public BuiltinType getType() {
         return new EntityType();
     }
@@ -35,6 +63,15 @@ public class EntityInstance extends BuiltinClass {
     @Override
     public int hashCode() {
         return Objects.hash(this.value);
+    }
+
+    @Override
+    public void setIndex(BuiltinClass index, BuiltinClass value) {
+        if (!index.instanceOf(new StringType())) {
+            throw ErrorHolder.cannotIndexTypeByType(this.getType(), index.getType());
+        }
+
+        ((EntityAccessor) this.value).sculk$writeCustomData(index.toString(), value.toNbtElement());
     }
 
     @Override